backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.26.0
redis==5.0.1
python-telegram-bot==21.0.1
openai==1.12.0
aiohttp==3.9.3
python-multipart==0.0.9
